Renamed folder 'enemies' to 'enemy_sprites'.

+++ src/entities/enemy_arrays.h
@@ -1,5 +1,5 @@
 #include "../images/enemy_sprites.h"
 
 const int enemy_array[] = {
-    300, 300, 25, 30, 25, alien
+    300, 300, 25, 30, 25, (int) alien
 };
